login&logout, Auth&AuthUser annotation

+++ src/main/java/com/cafe24/jblog/controller/UserController.java
@@ -13,16 +13,19 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import com.cafe24.jblog.service.UserService;
 import com.cafe24.jblog.vo.UserVo;
+import com.cafe24.security.Auth;
+import com.cafe24.security.AuthUser;
 
 @Controller
 @RequestMapping("/user")
 public class UserController {
 	@Autowired
 	private UserService userService;
-
+	
 	@RequestMapping(value = "/join", method = RequestMethod.GET)
 	public String join(@ModelAttribute UserVo userVo) {
 		return "user/join";
@@ -47,8 +50,20 @@ public class UserController {
 		return "user/joinsuccess";
 	}
 	
+	@RequestMapping(value = "/login", method = RequestMethod.GET)
+	public String login(@RequestParam(value = "result", required = false) String result,
+						Model model) {
+		model.addAttribute("result", result);
+		return "user/login";
+	}
+	
+	@Auth
 	@RequestMapping(value = "/delete/{id}")
-	public String delete(@PathVariable(value = "id") String id) {
+	public String delete(@AuthUser UserVo authUser, 
+						@PathVariable(value = "id") String id) {
+		if(authUser.getId()!=id) {
+			return "redirect:/";
+		}
 		userService.deleteUser(id);
 		return "redirect:/user";
 	}
